Make Connection, not RH instance
In the RHEV credentials form, we were previously attempting a connection
to the RH to make sure the credentials were valid. This got changed to
instantiating a RH model object, which threw an error and didn't
actually test the connection. Going back to trying to make a connection,
just with the new way of doing it with the new SDK.

[#157328361]