Access the sub-class specific attributes for OSBAs in the templates views.
When querying osbas from the resource handler, we only get an instance of OSBuildAttribute, which doesn't have access to sub-class fields like total_disk_size. Any resource handlers doing that query for templates to show on the view, it would show total_disk_size as N/A. You can call cast() on these instances (which I'm doing for LibCloudHandler) but the faster and better thing is to query directly from the submodels where we're already doing it from a tech-specific module.

[#158307680]