Conditionally show the Networks table column for total disk size dependening on whether the image class has the attribute.
Only some image classes have a total_disk_size attribute, so we shouldn't be showing this column when it isn't a real thing for the image. Conditionally show this column. This implementation is limited to our very NOT-DRY implementation of the images tab, which gets inserted manually by each technology's views. In the future, it would be much more DRY to clean those up and use the same view for all images.

[#158307680]